leave-management/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.1
SQLAlchemy==2.0.20
pydantic==2.11.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7  # Required for PostgreSQL on Render
